api/permissions.py
# axon_bbs/api/permissions.py
from rest_framework import permissions

class HasBoardAccess(permissions.BasePermission):
    """
    Custom permission to only allow users with the correct
    access level to see a message board.
    
    (This is a placeholder for a real permission system)
    """
    def has_object_permission(self, request, view, obj):
        # For now, we'll allow any authenticated user to see any board.
        # A real implementation would check obj.required_access_level
        # against the request.user.access_level.
        return request.user and request.user.is_authenticated

class HasFileAreaViewAccess(permissions.BasePermission):
    """
    Custom permission for viewing files in a FileArea.
    (Placeholder)
    """
    def has_object_permission(self, request, view, obj):
        # Similar to boards, allow any authenticated user for now.
        return request.user and request.user.is_authenticated

class HasFileAreaUploadAccess(permissions.BasePermission):
    """
    Custom permission for uploading files to a FileArea.
    (Placeholder)
    """
    def has_object_permission(self, request, view, obj):
        # Similar to boards, allow any authenticated user for now.
        return request.user and request.user.is_authenticated

